requirements.txt
python-socketio==5.11.4
uvicorn[standard]==0.44.0
